+++ hemocell.h
@@ -24,6 +24,7 @@
 #include "voxelizeDomain.h"
 #include "meshGeneratingFunctions.h"
 #include "loadBalancer.h"
+#include "preInlet.h"
 
 /* MECHANICS */
 #include "cellMechanics.h"